+++ static/src/components/table/table.ts
@@ -1,3 +1,3 @@
-import * as Tablesaw from "tablesaw";
-
+// import * as Tablesaw from "tablesaw";
+import * as Tablesaw from "tablesaw/dist/stackonly/tablesaw.stackonly.jquery";
 export default Tablesaw;